Initial implementation.
All functionality defined within a windows platform tag in plugin.xml, because
on other platforms, the compatibility library is not needed.

The library is not included in this plugin repository, but it is cloned by a
hook before the plugin is installed for windows platform. The clone is done
before plugin is installed so that the library file is present when it is
being copied onto the windows platform directory.

The library is included with a script element as the first child of the head
element so that the compatibility library is loaded before any other scripts
are evaluated. If it was included using Cordova's normal plugin script
inclusion, some scripts may have already tried to use incompatible functions.